Made script t0/t1/t2 compatible
